--- config/utils/index.ts
@@ -1,5 +1,0 @@
-import getCSSUses from './getCSSUses';
-
-export {
-	getCSSUses
-};
