+++ package.json
@@ -17,8 +17,8 @@
     "lint:ts": "tslint -p .",
     "lint": "npm-run-all lint:*",
     "prebuild": "rimraf ./dist/* ./types/*",
-    "build": "cross-env TS_NODE_COMPILER_OPTIONS={\\\"module\\\":\\\"commonjs\\\"} webpack --config ./config/webpack.config.ts",
-    "dev": "cross-env TS_NODE_COMPILER_OPTIONS={\\\"module\\\":\\\"commonjs\\\"} webpack-dev-server --config ./config/webpack.config.dev.ts --color",
+    "build": "vue-pack",
+    "dev": "vue-pack dev",
     "sync:db": "ts-node -O {\\\"module\\\":\\\"commonjs\\\"} ./scripts/sync-db.ts",
     "sync": "npm-run-all sync:*",
     "test:watch": "npm run lint && karma start --no-single-run",
@@ -50,71 +50,55 @@
     "vue": ">=2.4.0"
   },
   "devDependencies": {
-    "@types/chai": "^4.0.3",
-    "@types/copy-webpack-plugin": "^4.0.0",
-    "@types/html-webpack-plugin": "^3.0.0",
-    "@types/mini-css-extract-plugin": "^0.2.0",
-    "@types/mocha": "^5.2.0",
+    "@gluons/vue-pack": "^1.10.0",
+    "@gluons/vue-pack-banner-plugin": "^1.10.0",
+    "@gluons/vue-pack-cli": "^1.10.0",
+    "@gluons/vue-pack-splitchunks-plugin": "^1.10.0",
+    "@types/chai": "^4.1.5",
+    "@types/copy-webpack-plugin": "^4.4.2",
+    "@types/mocha": "^5.2.5",
     "@types/node": "8",
-    "@types/webpack": "^4.1.6",
-    "@types/webpack-merge": "^4.1.3",
-    "@vue/test-utils": "^1.0.0-beta.16",
-    "autoprefixer": "^9.0.0",
+    "@types/webpack": "^4.4.12",
+    "@types/webpack-node-externals": "^1.6.3",
+    "@vue/test-utils": "^1.0.0-beta.25",
     "babel-eslint": "^9.0.0",
     "browserslist-config-vue": "^1.0.3",
     "chai": "^4.1.1",
     "clipboardy": "^1.2.3",
-    "copy-webpack-plugin": "^4.0.1",
-    "cross-env": "^5.0.5",
-    "css-loader": "^0.28.4",
-    "ejs-loader": "^0.3.1",
-    "eslint": "^4.4.1",
+    "copy-webpack-plugin": "^4.5.2",
+    "cross-env": "^5.2.0",
+    "css-loader": "^1.0.0",
+    "eslint": "^5.6.0",
     "eslint-config-gluons": "^4.0.0",
-    "eslint-plugin-vue": "^4.5.0",
-    "fork-ts-checker-webpack-plugin": "^0.4.1",
-    "friendly-errors-webpack-plugin": "^1.7.0",
-    "html-webpack-plugin": "^3.2.0",
-    "karma": "^2.0.2",
+    "eslint-plugin-vue": "^4.7.1",
+    "karma": "^3.0.0",
     "karma-firefox-launcher": "^1.0.1",
     "karma-mocha": "^1.3.0",
     "karma-mocha-reporter": "^2.2.3",
     "karma-sourcemap-loader": "^0.3.7",
     "karma-webpack": "^4.0.0-beta.0",
-    "mini-css-extract-plugin": "^0.4.0",
-    "mocha": "^5.1.1",
-    "node-sass": "^4.5.3",
+    "mocha": "^5.2.0",
+    "node-sass": "^4.9.3",
     "npm-run-all": "^4.0.2",
-    "post-compile-webpack-plugin": "^0.1.2",
     "postcss-loader": "^3.0.0",
     "pug": "^2.0.3",
     "pug-plain-loader": "^1.0.0",
     "rimraf": "^2.6.1",
-    "sass-loader": "^7.0.1",
+    "sass-loader": "^7.1.0",
     "terminal-link": "^1.1.0",
-    "ts-loader": "^4.2.0",
-    "ts-node": "^7.0.0",
-    "tslint": "^5.6.0",
-    "tslint-config-gluons": "^2.0.0",
-    "typescript": "^3.0.0",
-    "vue": "^2.5.5",
+    "ts-node": "^7.0.1",
+    "tslint": "^5.11.0",
+    "tslint-config-gluons": "^2.0.3",
+    "typescript": "^3.0.3",
+    "vue": "^2.5.17",
     "vue-class-component": "^6.0.0",
-    "vue-loader": "^15.0.9",
-    "vue-template-compiler": "^2.5.5",
-    "webpack": "^4.7.0",
-    "webpack-cli": "^3.0.0",
-    "webpack-dev-server": "^3.1.4",
-    "webpack-merge": "^4.1.0",
-    "webpack-node-externals": "^1.7.2",
-    "webpack-stylish": "^0.1.8",
-    "webpackbar": "^2.6.1"
+    "vue-loader": "^15.4.2",
+    "vue-template-compiler": "^2.5.17",
+    "webpack": "^4.19.1",
+    "webpack-node-externals": "^1.7.2"
   },
   "dependencies": {
     "array-filter": "^1.0.0",
     "leven": "^2.1.0"
-  },
-  "resolutions": {
-    "@types/webpack": "^4.1.6",
-    "css-loader/cssnano/autoprefixer": "^8.4.1",
-    "css-loader/**/browserslist": "^3.2.7"
   }
 }
