+++ tsconfig.json
@@ -1,21 +1,21 @@
 {
 	"compilerOptions": {
-		"noUnusedLocals": true,
-		"noUnusedParameters": true,
+		"module": "es2015",
+		"moduleResolution": "node",
+		"target": "es5",
 		"lib": [
 			"dom",
 			"es2015"
 		],
-		"target": "es5",
-		"module": "es2015",
-		"moduleResolution": "node",
 		"esModuleInterop": true,
 		"experimentalDecorators": true,
-		"sourceMap": true
+		"sourceMap": true,
+		"noUnusedLocals": true,
+		"noUnusedParameters": true,
+		"declaration": true,
+		"declarationDir": "types"
 	},
 	"include": [
-		"./src/**/*",
-		"./dev/**/*",
-		"./test/**/*"
+		"src"
 	]
 }
