--- tsconfig.prod.json
@@ -1,11 +1,0 @@
-{
-	"extends": "./tsconfig",
-	"compilerOptions": {
-		"declaration": true,
-		"declarationDir": "types"
-	},
-	"exclude": [
-		"./dev/**/*",
-		"./test/**/*"
-	]
-}
